+++ omicverse/single/_batch.py
@@ -5,7 +5,8 @@ import anndata
 
 
 def batch_correction(adata:anndata.AnnData,batch_key:str,
-                     methods:str,n_pcs:int=50,**kwargs)->anndata.AnnData:
+                     use_rep='scaled|original|X_pca',
+                     methods:str='harmony',n_pcs:int=50,**kwargs)->anndata.AnnData:
     """
     Batch correction for single-cell data
 
@@ -35,7 +36,7 @@ def batch_correction(adata:anndata.AnnData,batch_key:str,
         adata3=adata.copy()
         scale(adata3)
         pca(adata3,layer='scaled',n_pcs=n_pcs)
-        sc.external.pp.harmony_integrate(adata3, batch_key,basis='scaled|original|X_pca',*kwargs)
+        sc.external.pp.harmony_integrate(adata3, batch_key,basis=use_rep,**kwargs)
         adata.obsm['X_harmony']=adata3.obsm['X_pca_harmony'].copy()
         return adata3
     elif methods=='combat':
@@ -43,7 +44,7 @@ def batch_correction(adata:anndata.AnnData,batch_key:str,
         sc.pp.combat(adata2, key=batch_key,*kwargs)
         scale(adata2)
         pca(adata2,layer='scaled',n_pcs=n_pcs)
-        adata2.obsm['X_combat']=adata2.obsm['scaled|original|X_pca'].copy()
+        adata2.obsm['X_combat']=adata2.obsm[use_rep].copy()
         adata.obsm['X_combat']=adata2.obsm['X_combat'].copy()
         return adata2
     elif methods=='scanorama':
@@ -68,7 +69,7 @@ def batch_correction(adata:anndata.AnnData,batch_key:str,
         adatas = list(alldata2.values())
         
         # run scanorama.integrate
-        scanorama.integrate_scanpy(adatas, dimred = n_pcs,*kwargs)
+        scanorama.integrate_scanpy(adatas, dimred = n_pcs,**kwargs)
         scanorama_int = [ad.obsm['X_scanorama'] for ad in adatas]
 
         # make into one matrix.